Remove the UI environment variable printout
In case this started working suddenly, it would not be good to be displaying these, so I only left the ones that are okay to be publicly shared. In the API route it matters less as it would only go to the log, but I removed it still out of an abudance of caution.